+++ R/dims.R
@@ -49,7 +49,6 @@ dims <- function(...,list.data,keep.names=TRUE,as.fun=NULL,keepNames){
 
     if(is.data.table(dots[[1]]) && is.null(as.fun)) as.fun <- "data.table"
     as.fun <- NMdataDecideOption("as.fun",as.fun)
-    
 
 
     dt.dims <- rbindlist(lapply(1:ndots,function(n) data.table(data=names.dots[n],nrows=nrow(dots[[n]]),ncols=ncol(dots[[n]]))))